+++ Runtime/System/Splines/MeshDeformer/CurveMeshDeformerPoint.cs
@@ -6,8 +6,10 @@ namespace Nazio_LT.Tools.Core
     public struct CurveMeshDeformerPointSettings
     {
         [SerializeField] private float distance;
+        [SerializeField] private bool genLast;
 
         public float Distance => distance > 0 ? distance : NMath.EPSILON;
+        public bool GenLast => genLast;
     }
 
     public class CurveMeshDeformerPoint : CurveMeshDeformer
@@ -21,22 +23,29 @@ namespace Nazio_LT.Tools.Core
 
         public override void Generate()
         {
-            int _objToPlace = (int)(settings.curve.curveLength / placementSettings.Distance);
+            Curve.Update();
+
+            int _objToPlace = (int)(settings.curve.curveLength / placementSettings.Distance) + 1;
             float _factor = 1f / (float)(_objToPlace - 1);
+
+            Debug.Log(settings.curve.curveLength + " : " + _objToPlace);
 
             for (var i = 0; i < _objToPlace; i++)
             {
-                PlaceMesh(i * _factor);
+                Debug.Log(i * placementSettings.Distance + " : " + i);
+                PlaceMesh(i * placementSettings.Distance);
             }
+
+            if (placementSettings.GenLast) PlaceMesh(settings.curve.curveLength);
         }
 
-        private void PlaceMesh(float _t)
+        private void PlaceMesh(float _dst)
         {
             GameObject _sub = CreateSubMesh(transform, out MeshRenderer _mr, out MeshFilter _mf);
 
             _mf.mesh = settings.meshToDeform;
 
-            _sub.transform.position = Curve.ComputePointUniform(_t, false);
+            _sub.transform.position = Curve.ComputePointDistance(_dst);
             _mr.material = Material;
         }
     }